requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
google-generativeai==0.3.2
duckduckgo-search==3.9.6
slowapi==0.1.9
pydantic==2.5.0
markdown==3.5.1
beautifulsoup4==4.12.2
lxml==4.9.3